Create Challenge Project: Credit Card Checker
I had some issues with the 'validateCred ()'method, as I initially forgot to include 'continue' to control the flow within the for loop. Prior to that, some of the elements in the array were passed through two separate cycles of multiplication. After that, it was somewhat smooth sailing. I felt that I may have been inefficient with how I constructed my 'idInvalideCardCompanies()' function. Feedback is welcome.